fix minor bug on mem_align, add more tests about mem_align

+++ src/sm/mem_align/mem_align_executor.hpp
@@ -17,7 +17,7 @@ public:
     mpz_class v;
     mpz_class w0;
     mpz_class w1;
-    uint8_t mode;
+    uint64_t mode;
     uint8_t wr;
 };
